+++ rioxarray/exceptions.py
@@ -12,20 +12,24 @@ class NoDataInBounds(RioXarrayError):
     """This is for when there are no data in the bounds for clipping a raster."""
 
 
-class OneDimensionalRaster(RioXarrayError):
-    """This is an error when you have a 1 dimensional raster."""
-
-
 class SingleVariableDataset(RioXarrayError):
     """This is for when you have a dataset with a single variable."""
 
 
-class TooManyDimensions(RioXarrayError):
+class DimensionError(RioXarrayError):
     """This is raised when there are more dimensions than is supported by the method"""
 
 
-class InvalidDimensionOrder(RioXarrayError):
+class TooManyDimensions(DimensionError):
+    """This is raised when there are more dimensions than is supported by the method"""
+
+
+class InvalidDimensionOrder(DimensionError):
     """This is raised when there the dimensions are not ordered correctly."""
+
+
+class OneDimensionalRaster(DimensionError):
+    """This is an error when you have a 1 dimensional raster."""
 
 
 class MissingCRS(RioXarrayError):